// src/domain/models/RazaGato.ts
export interface RazaGato {
  id: string;
  nombre: string;
  origen: string;
  temperamento: string;
  descripcion: string;
  esperanzaVida: string;
  pesoMetrico: string;
  adaptabilidad: number;
  inteligencia: number;
  nivelAfecto: number;
  imagenUrl: string | null;
  wikipediaUrl: string | null;
}